+++ prisma/migrations/20260813153419_external_account_username/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "external_accounts" ADD COLUMN     "username" TEXT;